+++ src/Symfony/Component/Cache/Adapter/PdoAdapter.php
@@ -12,10 +12,6 @@
 namespace Symfony\Component\Cache\Adapter;
 
 use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Schema\Schema;
-use Psr\Cache\CacheItemInterface;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Cache\CacheItem;
 use Symfony\Component\Cache\Exception\InvalidArgumentException;
 use Symfony\Component\Cache\Marshaller\DefaultMarshaller;
 use Symfony\Component\Cache\Marshaller\MarshallerInterface;
@@ -40,8 +36,6 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
     private mixed $connectionOptions = [];
     private string $namespace;
 
-    private $dbalAdapter;
-
     /**
      * You can either pass an existing database connection as PDO instance or
      * a DSN string that will be used to lazy-connect to the database when the
@@ -61,13 +55,10 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
      * @throws InvalidArgumentException When PDO error mode is not PDO::ERRMODE_EXCEPTION
      * @throws InvalidArgumentException When namespace contains invalid characters
      */
-    public function __construct(\PDO|Connection|string $connOrDsn, string $namespace = '', int $defaultLifetime = 0, array $options = [], MarshallerInterface $marshaller = null)
+    public function __construct(\PDO|string $connOrDsn, string $namespace = '', int $defaultLifetime = 0, array $options = [], MarshallerInterface $marshaller = null)
     {
-        if ($connOrDsn instanceof Connection || (\is_string($connOrDsn) && str_contains($connOrDsn, '://'))) {
-            trigger_deprecation('symfony/cache', '5.4', 'Usage of a DBAL Connection with "%s" is deprecated and will be removed in symfony 6.0. Use "%s" instead.', __CLASS__, DoctrineDbalAdapter::class);
-            $this->dbalAdapter = new DoctrineDbalAdapter($connOrDsn, $namespace, $defaultLifetime, $options, $marshaller);
-
-            return;
+        if (\is_string($connOrDsn) && str_contains($connOrDsn, '://')) {
+            throw new InvalidArgumentException(sprintf('Usage of Doctrine DBAL URL with "%s" is not supported. Use a PDO DSN or "%s" instead. Got "%s".', __CLASS__, DoctrineDbalAdapter::class, $connOrDsn));
         }
 
         if (isset($namespace[0]) && preg_match('#[^-+.A-Za-z0-9]#', $namespace, $match)) {
@@ -80,10 +71,8 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
             }
 
             $this->conn = $connOrDsn;
-        } elseif (\is_string($connOrDsn)) {
-            $this->dsn = $connOrDsn;
         } else {
-            throw new InvalidArgumentException(sprintf('"%s" requires PDO or Doctrine\DBAL\Connection instance or DSN string as first argument, "%s" given.', __CLASS__, get_debug_type($connOrDsn)));
+            $this->dsn = $connOrDsn;
         }
 
         $this->table = $options['db_table'] ?? $this->table;
@@ -101,166 +90,6 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
     }
 
     /**
-     * {@inheritDoc}
-     */
-    public function getItem($key): CacheItem
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->getItem($key);
-        }
-
-        return parent::getItem($key);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getItems(array $keys = []): iterable
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->getItems($keys);
-        }
-
-        return parent::getItems($keys);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function hasItem($key): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->hasItem($key);
-        }
-
-        return parent::hasItem($key);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function deleteItem($key): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->deleteItem($key);
-        }
-
-        return parent::deleteItem($key);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function deleteItems(array $keys): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->deleteItems($keys);
-        }
-
-        return parent::deleteItems($keys);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function clear(string $prefix = ''): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->clear($prefix);
-        }
-
-        return parent::clear($prefix);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function get(string $key, callable $callback, float $beta = null, array &$metadata = null): mixed
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->get($key, $callback, $beta, $metadata);
-        }
-
-        return parent::get($key, $callback, $beta, $metadata);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function delete(string $key): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->delete($key);
-        }
-
-        return parent::delete($key);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function save(CacheItemInterface $item): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->save($item);
-        }
-
-        return parent::save($item);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function saveDeferred(CacheItemInterface $item): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->saveDeferred($item);
-        }
-
-        return parent::saveDeferred($item);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setLogger(LoggerInterface $logger): void
-    {
-        if (isset($this->dbalAdapter)) {
-            $this->dbalAdapter->setLogger($logger);
-
-            return;
-        }
-
-        parent::setLogger($logger);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function commit(): bool
-    {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->commit();
-        }
-
-        return parent::commit();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function reset(): void
-    {
-        if (isset($this->dbalAdapter)) {
-            $this->dbalAdapter->reset();
-
-            return;
-        }
-
-        parent::reset();
-    }
-
-    /**
      * Creates the table to store cache items which can be called once for setup.
      *
      * Cache ID are saved in a column of maximum length 255. Cache data is
@@ -271,12 +100,6 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
      */
     public function createTable()
     {
-        if (isset($this->dbalAdapter)) {
-            $this->dbalAdapter->createTable();
-
-            return;
-        }
-
         // connect if we are not yet
         $conn = $this->getConnection();
 
@@ -309,26 +132,10 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
     }
 
     /**
-     * Adds the Table to the Schema if the adapter uses this Connection.
-     *
-     * @deprecated since symfony/cache 5.4 use DoctrineDbalAdapter instead
-     */
-    public function configureSchema(Schema $schema, Connection $forConnection): void
-    {
-        if (isset($this->dbalAdapter)) {
-            $this->dbalAdapter->configureSchema($schema, $forConnection);
-        }
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function prune(): bool
     {
-        if (isset($this->dbalAdapter)) {
-            return $this->dbalAdapter->prune();
-        }
-
         $deleteSql = "DELETE FROM $this->table WHERE $this->lifetimeCol + $this->timeCol <= :time";
 
         if ('' !== $this->namespace) {